Move assets to assets folder
