+++ assets/src/scripts/test/fixtures/version_from_scratch.json
