+++ assets/src/scripts/test/fixtures/version_with_complete_searches.json
