+++ assets/src/scripts/test/fixtures/version_with_no_searches.json
