+++ assets/src/scripts/test/fixtures/version_with_some_searches.json
